requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
praw>=7.7.0
